+++ pyrestore/__init__.py
@@ -1,4 +1,6 @@
 from .pyrestore import Pyrestore, FieldValue, FirestoreQuery, Batch
 from .firebase import FirebaseManager
+from .rulebuilder import RuleBuilder
+from .storage import Storage
 
-__all__ = ["Pyrestore", "FieldValue", "FirestoreQuery", "Batch", "FirebaseManager"]
+__all__ = ["Pyrestore", "RuleBuilder", "FieldValue", "FirestoreQuery", "Storage", "Batch", "FirebaseManager"]