+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the `pyrestore` project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2026-07-23
+
+### Changed
+- **Updated README**: Added descriptive text for how to use `Storage` the synchronous and asynchronous methods.
+
+### Added
+- **Storage Module**: Created a low-level storage module for accessing Firebase storage, with high-level functions
+added to `FirebaseManager()`.
+- **RuleBuilder**: Create an easy to use rule-builder tool for those who are new to or struggle with writing firebase rules
+- **RuleBuilderCheatsheet.md**: Created a separate `.md` file for how to use the new `RuleBuilder()` tool.
 ---
 
 ## [1.0.2] - 2026-07-23